Improved customer search by name
Now searches customers by name after the following rules: if :name
contains only one word, customers are search by this name both in
first and last names. If :name consists of more than one word, the
last word is interpreted as last name and the other words are
considered to be first names.